requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
mysqlclient==2.2.7
